app/requirements.txt
TA-Lib
python-binance
numpy
flask
datetime
